created comment controller
